+++ apps/framework/src/runtime/database/process/registry.ts
@@ -3,7 +3,6 @@ import { coreDatabaseModule } from "../../../../../core/src/database-module.js"
 import { cxappDatabaseModule } from "../../../../../cxapp/src/database-module.js"
 import { ecommerceDatabaseModule } from "../../../../../ecommerce/src/database-module.js"
 import { frappeDatabaseModule } from "../../../../../frappe/src/database-module.js"
-import { zetroDatabaseModule } from "../../../../../zetro/src/database-module.js"
 
 import { frameworkDatabaseMigrations } from "./migrations/index.js"
 import {
@@ -28,7 +27,6 @@ export const registeredDatabaseModules: AppDatabaseModule[] = [
   billingDatabaseModule,
   ecommerceDatabaseModule,
   frappeDatabaseModule,
-  zetroDatabaseModule,
 ].sort((left, right) => left.order - right.order || left.appId.localeCompare(right.appId))
 
 export function listRegisteredDatabaseMigrations() {